fix(cognitive-load): return per-person ratio for service load dimension
The service load dimension was returning the raw service count as its
value but assessing the level based on the services-per-person ratio.
This mismatch made the UI confusing: a team with 5 services and 5
people showed "5 low" next to threshold text ">3 high".

Now LoadDimension.Value is float64 and assessServiceLoadCfg returns
the actual ratio (e.g. 1.0 for 5/5) so the displayed value matches
the thresholds. All consumers updated: view handlers, AI prompt data,
CLI formatter, and signal suggestions.

Made-with: Cursor

+++ backend/internal/adapter/handler/ai.go
@@ -578,13 +578,13 @@ func buildAIPromptData(m *entity.UNMModel, userQuestion string, h *Handler) map[
 		Team               string
 		OverallLevel       string
 		DomainSpread       string
-		DomainSpreadVal    int
+		DomainSpreadVal    float64
 		ServiceLoad        string
-		ServiceLoadVal     int
+		ServiceLoadVal     float64
 		InteractionLoad    string
-		InteractionLoadVal int
+		InteractionLoadVal float64
 		DependencyLoad     string
-		DependencyLoadVal  int
+		DependencyLoadVal  float64
 		ServiceCount       int
 		CapabilityCount    int
 		TeamSize           int
